// toolkit/packages/pancake-uikit/src/widgets/Menu/components/MenuEntry.tsx
import React from "react";
import styled, { keyframes, DefaultTheme } from "styled-components";
import { Text } from "../../../components/Text";
import { Colors } from "../../../theme/types";
import { MENU_ENTRY_HEIGHT } from "../config";

export interface Props {
  secondary?: boolean;
  isActive?: boolean;
  theme: DefaultTheme;
}

const rainbowAnimation = keyframes`
  0%,
  100% {
    background-position: 0 0;
  }
  50% {
    background-position: 100% 0;
  }
`;

const LinkLabel = styled.div<{ isPushed: boolean }>`
  color: ${({ isPushed, theme }) => (isPushed ? theme.colors.menuText : "transparent")};
  transition: color 0.4s;
  flex-grow: 1;
`;

const MenuEntry = styled.div<Props>`
  cursor: pointer;
  display: flex;
  align-items: center;
  height: ${MENU_ENTRY_HEIGHT}px;
  padding: ${({ secondary }) => (secondary ? "0 32px" : "0 25px")};
  font-size: ${({ secondary }) => (secondary ? "12px" : "14px")};
  background-color: ${({ secondary, theme }) => (secondary ? theme.colors.background : "transparent")};
  color: ${({ theme }) => theme.colors.menuText};
  // box-shadow: ${({ isActive, theme }) => (isActive ? `rgb(183, 183, 183) 0px -3px 0px inset` : "none")};
  background: ${({ isActive, theme }) => (isActive ? 'rgba(57,208,216,.1)' : "none")};
  border-radius: ${({ isActive, theme }) => (isActive ? '0.75rem' : "none")};
  padding-left: ${({ isActive, theme }) => (isActive ? '1rem' : "none")};
  padding-right: ${({ isActive, theme }) => (isActive ? '1rem' : "none")};
  margin-left: ${({ isActive, theme }) => (isActive ? '7px' : "none")};
  margin-right: ${({ isActive, theme }) => (isActive ? '10px' : "none")};


  a {
    display: flex;
    align-items: center;
    width: 100%;
    height: 100%;
  }

  svg {
    fill: ${({ theme }) => theme.colors.textSubtle};
  }

  &:hover {
    background-color: ${({ theme }) => 'rgba(57,208,216,.1)'};
    border-radius: ${({ theme }) => '0.75rem'};
  }

  // Safari fix
  flex-shrink: 0;

  &.rainbow {
    background-clip: text;
    animation: ${rainbowAnimation} 3s ease-in-out infinite;
    background: ${({ theme }) => theme.colors.gradients.bubblegum};
    background-size: 400% 100%;
  }
`;
MenuEntry.defaultProps = {
  secondary: false,
  isActive: false,
};

const LinkStatus = styled(Text)<{ color: keyof Colors }>`
  border-radius: ${({ theme }) => theme.radii.default};
  padding: 0 8px;
  border: 2px solid;
  border-color: ${({ theme, color }) => theme.colors[color]};
  box-shadow: none;
  color: ${({ theme, color }) => theme.colors[color]};
  margin-left: 8px;
`;

const LinkLabelMemo = React.memo(LinkLabel, (prev, next) => prev.isPushed === next.isPushed);

export { MenuEntry, LinkStatus, LinkLabelMemo as LinkLabel };
